+++ examples/repo_dag.rs
@@ -1,6 +1,3 @@
-use std::future::IntoFuture;
-
-use futures::join;
 use libipld::ipld;
 use rust_ipfs::dag::IpldDag;
 use rust_ipfs::repo::Repo;
@@ -11,25 +8,22 @@ async fn main() -> anyhow::Result<()> {
     tracing_subscriber::fmt::init();
 
     // Initialize the repo
-    let repo = Repo::new_memory();
+    let repo = Repo::new_memory(None);
     let dag = IpldDag::from(repo.clone());
 
-    // Create a DAG
-    let f1 = dag.put_dag(ipld!("block1")).into_future();
-    let f2 = dag.put_dag(ipld!("block2")).into_future();
-    let (res1, res2) = join!(f1, f2);
-    let root = ipld!([res1?, res2?]);
+    let cid1: libipld::cid::CidGeneric<64> = dag.put_dag(ipld!("block1")).await?;
+    let cid2 = dag.put_dag(ipld!("block2")).await?;
+    let root = ipld!([cid1, cid2]);
     let cid = dag.put_dag(root).await?;
     let path = IpfsPath::from(cid);
 
     // Query the DAG
     let path1 = path.sub_path("0")?;
     let path2 = path.sub_path("1")?;
-    let f1 = dag.get_dag(path1).into_future();
-    let f2 = dag.get_dag(path2).into_future();
-    let (res1, res2) = join!(f1, f2);
-    println!("Received block with contents: {:?}", res1?);
-    println!("Received block with contents: {:?}", res2?);
+    let block1 = dag.get_dag(path1).await?;
+    let block2 = dag.get_dag(path2).await?;
+    println!("Received block with contents: {:?}", block1);
+    println!("Received block with contents: {:?}", block2);
 
     repo.shutdown();
     Ok(())