+++ src/repo/blockstore/mod.rs
@@ -1,3 +1,4 @@
+use futures::stream::BoxStream;
 use libipld::Cid;
 
 use crate::{Block, Channel};
@@ -16,6 +17,13 @@ pub(crate) enum RepoBlockCommand {
         cid: Cid,
         response: Channel<Option<Block>>,
     },
+    Size {
+        cid: Vec<Cid>,
+        response: Channel<Option<usize>>,
+    },
+    TotalSize {
+        response: Channel<usize>,
+    },
     PutBlock {
         block: Block,
         response: Channel<(Cid, BlockPut)>,
@@ -23,6 +31,10 @@ pub(crate) enum RepoBlockCommand {
     Remove {
         cid: Cid,
         response: Channel<Result<BlockRm, BlockRmError>>,
+    },
+    Cleanup {
+        refs: BoxStream<'static, Cid>,
+        response: Channel<Vec<Cid>>,
     },
     List {
         response: Channel<Vec<Cid>>,