+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # next [unrelease]
 - chore: Implements metrics [PR XXX](https://github.com/dariusc93/rust-ipfs/pull/XXX)
+- feat: Implements garbage collector. [PR 120](https://github.com/dariusc93/rust-ipfs/pull/120)
 - refactor: Revise peerbook and move connect/disconnect operations to main task [PR 119](https://github.com/dariusc93/rust-ipfs/pull/119)
 
 # 0.7.1